backend/requirements.txt
# Core
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.4.0
python-dotenv==1.0.0

# LLM & AI
openai==1.3.0
google-generativeai==0.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Database
redis==5.0.0
sqlalchemy==2.0.23

# Financial APIs
plaid-python==15.0.0
stripe==7.0.0

# Utilities
httpx==0.25.0
pillow==10.1.0
websockets==12.0
python-multipart==0.0.6

# Web Scraping (for agents)
beautifulsoup4==4.12.2
lxml==4.9.3

# Gmail API (optional - for BountyHunter1)
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
